+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to Burrow will be documented here.
 
 ## [Unreleased]
 
+## [0.3.6-beta] - 2026-08-05
+
+### Added
+
+- Extended the Home and Store footer to nested manual folders inside the configured library
+
+### Fixed
+
+- Anchored the folder and collection marker to the upper-left corner of the cover
+- Applied real rounded-corner masking to manual-folder artwork in Cover Grid and Cover List
+
 ## [0.3.5-beta] - 2026-08-05
 
 ### Added
@@ -129,7 +140,8 @@ All notable changes to Burrow will be documented here.
 
 - Initial unified Burrow library, Store, and reader-interface package
 
-[Unreleased]: https://github.com/richbeatty/burrow.koplugin/compare/v0.3.5-beta...HEAD
+[Unreleased]: https://github.com/richbeatty/burrow.koplugin/compare/v0.3.6-beta...HEAD
+[0.3.6-beta]: https://github.com/richbeatty/burrow.koplugin/releases/tag/v0.3.6-beta
 [0.3.5-beta]: https://github.com/richbeatty/burrow.koplugin/releases/tag/v0.3.5-beta
 [0.3.4-beta]: https://github.com/richbeatty/burrow.koplugin/releases/tag/v0.3.4-beta
 [0.3.3-beta]: https://github.com/richbeatty/burrow.koplugin/releases/tag/v0.3.3-beta